+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 BI indicates a backward incompatible change. Take caution when upgrading to a
 version with these. Your code will need to be updated to continue working.
 
+## 2.0.3
+
+ * fixed breaking change in 2.0.2 where context did not have uidb36 and token
+
 ## 2.0.2
 
  * fixed potentional security issue with leaking password reset tokens through HTTP Referer header